magazine-master-updated/magazine-master/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0b1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.10.3
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.2
importlib-resources==1.4.0
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
pypiwin32==223
python3-openid==3.1.0
pytz==2019.3
pywin32==227
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.4
zipp==3.1.0
